Don't add deopt points post-materialization
We were currently adding these for things we'd started to consider to
be irreplaceable, which was harmless but wasteful. As we prepare to do
partial escape analysis, however, it's a correctness issue that we do
not add such entries beyond the point where the object is materialized.
Otherwise, we'd potentially overwrite the object with another one with
state values and a different identity.