Keep track of inserted materialization transforms
So we can far more easily detect the case where we have a usage of a
materialized value and insert it into the list of aliases to write into
upon a materialization. This will also come in useful for fixing the SSA
form problems encountered when we materialize an object on two different
sides of a branch. No functional changes are intended as a result of
this; it's just a preparatory refactor.